Move glfw outta lib
